requirements.txt
fastapi
uvicorn[standard]
pyyaml
websockets
pytest
pytest-asyncio
httpx
